# [POC] Optimize performance and resources of ComposerRepository