I noticed that Composer always loads all of the package provider data into memory even if it could know exactly that this information is never going to be needed because it cannot resolve to valid set of dependencies.

The `Pool` already knew about the root requirements (I renamed the variable because `filterRequires` felt awkward) but not about the root conflicts, so I added that.

What happens now is that if you required for example `symfony/framework-bundle: ^4.3`, every version that does not match is not loaded into memory anymore, thus speeding up the whole process and reducing memory usage.
The same applies for conflicts.

In case you wonder what happens, if you add a conflict that strips all tags from a package, just try it out by e.g. configuring a conflict to everything lower than `symfony/framework-bundle` in version `5.0` (which doesn't exist):

```json
{
    "name": "my/project",
    "type": "project",
    "require": {
        "php": "^7.1",
        "symfony/framework-bundle": "^4.3"
    },
    "conflict": {
        "symfony/framework-bundle": "<5"
    }
}
```

Result:

```
  Problem 1
    - symfony/framework-bundle v4.3.0 conflicts with my/project[No version set (parsed as 1.0.0)].
    - symfony/framework-bundle v4.3.1 conflicts with my/project[No version set (parsed as 1.0.0)].
    - symfony/framework-bundle v4.3.2 conflicts with my/project[No version set (parsed as 1.0.0)].
    - symfony/framework-bundle v4.3.3 conflicts with my/project[No version set (parsed as 1.0.0)].
    - Installation request for my/project No version set (parsed as 1.0.0) -> satisfiable by my/project[No version set (parsed as 1.0.0)].
    - Installation request for symfony/framework-bundle ^4.3 -> satisfiable by symfony/framework-bundle[v4.3.0, v4.3.1, v4.3.2, v4.3.3].
```

So as you can see, it clearly states that it "conflicts with my/project", so the root composer.json 😊 

By adding more conflicts or better requires I can now considerably reduce memory usage and speed up the resolving process.