@nicolas-grekas 2019-08-13
It'd be interesting to have some numbers on real-world apps (flex disabled of course)
Then, I'll let Jordi and Nils be the judge, they're the expert here :)