@Toflar 2019-08-13
I‘ll have a look at the testsuite and provide some numbers when I find some more time to work on this :)