@alcohol 2019-08-14
For starters, you're missing a `return true` at the end of `isPackageAllowedByRoot`