@Seldaek 2019-08-16
TBH I'd rather also not do this right now, sounds like a last resort thing rather.. 

"Forcing" everyone to declare conflicts for all their transitive dependencies sucks. Plus the risk of seeing this practice proliferate all over the place in open source package, would end up having to load many more objects to solve the dependencies than we have now, as we need one conflict object per version of each package, this grows quickly and might even offset the gains I don't know.