@Seldaek 2019-08-16
To sum up, I have no doubt it'd be effective if well targeted etc, but if you look at the whole ecosystem it sounds like a net negative.