@Toflar 2019-08-19
The thing is, the solver actually seems to be pretty fast so the ruleset size doesn't seem to be the real problem. It's more the `ComposerRepository` which just loads a massive amount of JSON data into memory and for some reason lookups slow down (otherwise, why would it be that much faster?).

I understand your point of view saying that it will maybe trigger some bad practice solutions. But fact is, it's already being done. What Symfony Flex does is exactly that and people love it because they just don't care about all the `2.x` versions anymore.
I'm just saying that if we introduced that in Composer then Flex could continue to do the same as it does today by just adding conflicts instead of replacing the `ComposerRepository` class. You cannot have multiple plugins that do the same at the moment. If they added root conflicts, you could. 

But I really want to understand why exactly lookups seem to be slowed down that much. I sort of have an idea but I want to look at it more closely before I write reams again :P