@andrerom 2019-08-19
Tried this, and did not see any noticeable difference so far.

`composer update` memory usage on `ezsystems/ezpublish-kernel:7.5`
PHP: PHP 7.3.6 ( NTS ), Zend OPcache v7.3.6,  Xdebug v2.7.2

```
v1.9.0 --no-plugins:
[347.4MiB/14.44s] Memory usage: 347.4MiB (peak: 1619.91MiB), time: 14.44s

v1.9.0 with symfony flex (+ "extra.symfony.require: 3.4.x"):
[212.5MiB/3.26s] Memory usage: 212.49MiB (peak: 480.82MiB), time: 3.26s

With filter-root-conflicts-and-requires:
[347.5MiB/14.15s] Memory usage: 347.48MiB (peak: 1620.07MiB), time: 14.15s

With filter-root-conflicts-and-requires and conflict rules to rule out Symfony <3.4, >=3.5:
[346.7MiB/14.33s] Memory usage: 346.66MiB (peak: 1619.93MiB), time: 14.33s

Composer 2.0, NB! stops on conflict on "ocramius/package-versions * requires composer-plugin-api ^1.0", so not an apples to apples comparison:
[108.2MiB/12.96s] Memory usage: 108.2MiB (peak: 1374.8MiB), time: 12.96s 
```

@Toflar Feel free to message me on twitter to go over testing as I might have overlooked a way to trigger this. _But_ from a user view I have to agree with Jordi that having to define conflicts to rule out versions seems to not be very user friendly, it'll produce a lot of dependency issues once lib maintainers start using it.

@Seldaek That said I hope you agree this is a major/growing issue.
   _(FYI On PHP 5.6 we have x3, so around 4GiB, which we still need to maintain. And AFAICS within next 2 years numbers will probably grow 2-3 times across PHP versions, as size of package dependencies versions graph exponentially grows)_