@Toflar 2019-08-20
I've adjusted the implementation and here are some numbers. I've tested using the following `composer.json`:

```json
{
    "name": "my/contao-project",
    "type": "project",
    "require": {
        "php": "^5.6 || ^7.0",
        "contao/calendar-bundle": "^4.4",
        "contao/comments-bundle": "^4.4",
        "contao/faq-bundle": "^4.4",
        "contao/listing-bundle": "^4.4",
        "contao/manager-bundle": "4.4.*",
        "contao/news-bundle": "^4.4",
        "contao/newsletter-bundle": "^4.4"
    }
}
```

The command I run is `bin/composer update --dry-run --profile`.

* **master**
   Memory usage: 366.82MiB (peak: 1875.88MiB), time: 28.75s
* **with this PR**
   Memory usage: 365.03MiB (peak: 1873.79MiB), time: 25.21s

So there's an ever so slight improvement because of the root requirements filtered out. However, there are not too many tags in that particular case so gains are marginal.
Now let's add some conflicts:

```json
{
    "name": "my/contao-project",
    "type": "project",
    "require": {
        "php": "^5.6 || ^7.0",
        "contao/calendar-bundle": "^4.4",
        "contao/comments-bundle": "^4.4",
        "contao/faq-bundle": "^4.4",
        "contao/listing-bundle": "^4.4",
        "contao/manager-bundle": "4.4.*",
        "contao/news-bundle": "^4.4",
        "contao/newsletter-bundle": "^4.4"
    },
    "conflict": {
        "symfony/console": "<3.3",
        "symfony/filesystem": "<3.3",
        "symfony/framework-bundle": "<3.3",
        "symfony/http-foundation": "<3.3",
        "symfony/security": "<3.3",
        "symfony/yaml": "<3.3",
        "symfony/event-dispatcher": "<3.3",
        "symfony/options-resolver": "<3.3",
        "symfony/http-kernel": "<3.3",
        "doctrine/dbal": "<2.5"
    }
}
```

* **master**
   Memory usage: 366.82MiB (peak: 1876.58MiB), time: 27.98s
* **with this PR**
   Memory usage: 312.78MiB (peak: 1092.01MiB), time: 14.68s

So it is **considerably** faster and more memory efficient 😄 
Of course you could also move the conflicts to the requirements. It would work just the same and have the same effect (I did test that).

I've also updated the problem information. So if I conflict with my own requirements like so:

```json
{
    "name": "my/contao-project",
    "type": "project",
    "require": {
        "php": "^5.6 || ^7.0",
        "contao/calendar-bundle": "^4.4",
        "contao/comments-bundle": "^4.4",
        "contao/faq-bundle": "^4.4",
        "contao/listing-bundle": "^4.4",
        "contao/manager-bundle": "4.4.*",
        "contao/news-bundle": "^4.4",
        "contao/newsletter-bundle": "^4.4"
    },
    "conflict": {
        "contao/manager-bundle": "^4.4"
    }
}
```

I get this:

```
Your requirements could not be resolved to an installable set of packages.
  Problem 1
    - The requested package contao/manager-bundle 4.4.* was excluded by your own root composer.json "conflict" definition.
```

It also works for conflicts of transitive dependencies:

```json
{
    "name": "my/contao-project",
    "type": "project",
    "require": {
        "php": "^5.6 || ^7.0",
        "contao/calendar-bundle": "^4.4",
        "contao/comments-bundle": "^4.4",
        "contao/faq-bundle": "^4.4",
        "contao/listing-bundle": "^4.4",
        "contao/manager-bundle": "4.4.*",
        "contao/news-bundle": "^4.4",
        "contao/newsletter-bundle": "^4.4"
    },
    "conflict": {
        "doctrine/dbal": ">=2.5"
    }
}
```

This is what you get:

```
[...]
    - contao/core-bundle 4.4.1 requires doctrine/dbal ^2.5 -> conflicts with your root composer.json "conflict" definition.
    - contao/core-bundle 4.4.0 requires doctrine/dbal ^2.5 -> conflicts with your root composer.json "conflict" definition.
    - Installation request for contao/manager-bundle 4.4.* -> satisfiable by contao/manager-bundle[4.4.0, 4.4.1, 4.4.10, 4.4.11, 4.4.12, 4.4.13, 4.4.14, 4.4.15, 4.4.16, 4.4.17, 4.4.18, 4.4.19, 4.4.2, 4.4.20, 4.4.21, 4.4.22, 4.4.23, 4.4.24, 4.4.25, 4.4.26, 4.4.27, 4.4.28, 4.4.29, 4.4.3, 4.4.30, 4.4.31, 4.4.32, 4.4.33, 4.4.34, 4.4.35, 4.4.36, 4.4.37, 4.4.38, 4.4.39, 4.4.4, 4.4.40, 4.4.41, 4.4.42, 4.4.5, 4.4.6, 4.4.7, 4.4.8, 4.4.9].
```