@Toflar 2019-08-20
And here is a comparison for a typical Symfony application (without Flex) plus all the requirements of `api-pack` so API platform dependencies.
Note that **I did not specify any special conflicts**! The optimizations kick in by just having the root requirements filter out impossible candidates in that case:

```json
{
    "name": "my/symfony-project",
    "type": "project",
    "require": {
        "php": "^7.2.9",
        "ext-ctype": "*",
        "ext-iconv": "*",
        "symfony/console": "^4.3",
        "symfony/dotenv": "^4.3",
        "symfony/framework-bundle": "^4.3",
        "symfony/yaml": "^4.3",

        "api-platform/core": "^2.1",
        "doctrine/annotations": "^1.0",
        "doctrine/orm": "^2.4.5",
        "doctrine/doctrine-bundle": "^1.6",
        "nelmio/cors-bundle": "^1.5",
        "phpdocumentor/reflection-docblock": "^3.0 || ^4.0",
        "symfony/asset": "^4.3",
        "symfony/expression-language": "^4.3",
        "symfony/security-bundle": "^4.3",
        "symfony/twig-bundle": "^4.3",
        "symfony/validator": "^4.3"
    }
}
```


* **master**
    Memory usage: 273.58MiB (peak: 661.23MiB), time: 7.34s
* **with this PR**
    Memory usage: 247.97MiB (peak: 635.68MiB), time: 6.57s

I mean, it's not huge but it's 10% less memory used anyway, without finetuning anything.
And now if I were to analyze what packages it installs, I can optimize a bit more by adding a few conflicts (I don't need to specify thousands of packages, it kicks in pretty quickly):

```json
{
    "name": "my/symfony-project",
    "type": "project",
    "require": {
        "php": "^7.2.9",
        "ext-ctype": "*",
        "ext-iconv": "*",
        "symfony/console": "^4.3",
        "symfony/dotenv": "^4.3",
        "symfony/framework-bundle": "^4.3",
        "symfony/yaml": "^4.3",

        "api-platform/core": "^2.1",
        "doctrine/annotations": "^1.0",
        "doctrine/orm": "^2.4.5",
        "doctrine/doctrine-bundle": "^1.6",
        "nelmio/cors-bundle": "^1.5",
        "phpdocumentor/reflection-docblock": "^3.0 || ^4.0",
        "symfony/asset": "^4.3",
        "symfony/expression-language": "^4.3",
        "symfony/security-bundle": "^4.3",
        "symfony/twig-bundle": "^4.3",
        "symfony/validator": "^4.3"
    },
    "conflict": {
        "symfony/yaml": "<4.3",
        "symfony/serializer": "<4.3",
        "symfony/http-foundation": "<4.3",
        "symfony/filesystem": "<4.3",
        "symfony/cache": "<4.3",
        "symfony/config": "<4.3"
    }
}
```

* **with this PR**
    Memory usage: 236.28MiB (peak: 569.22MiB), time: 5.95s

And now I can go a bit crazy and add some more conflicts (or even better, only add them to the `composer.json` dynamically during the CI job) because my CI does not offer 570 MB of RAM:

```json
{
    "name": "my/symfony-project",
    "type": "project",
    "require": {
        "php": "^7.2.9",
        "ext-ctype": "*",
        "ext-iconv": "*",
        "symfony/console": "^4.3",
        "symfony/dotenv": "^4.3",
        "symfony/framework-bundle": "^4.3",
        "symfony/yaml": "^4.3",

        "api-platform/core": "^2.1",
        "doctrine/annotations": "^1.0",
        "doctrine/orm": "^2.4.5",
        "doctrine/doctrine-bundle": "^1.6",
        "nelmio/cors-bundle": "^1.5",
        "phpdocumentor/reflection-docblock": "^3.0 || ^4.0",
        "symfony/asset": "^4.3",
        "symfony/expression-language": "^4.3",
        "symfony/security-bundle": "^4.3",
        "symfony/twig-bundle": "^4.3",
        "symfony/validator": "^4.3"
    },
    "conflict": {
        "symfony/yaml": "<4.3",
        "symfony/serializer": "<4.3",
        "symfony/http-foundation": "<4.3",
        "symfony/filesystem": "<4.3",
        "symfony/cache": "<4.3",
        "symfony/config": "<4.3",
        "symfony/routing": "<4.3",
        "symfony/finder": "<4.3",
        "symfony/var-exporter": "<4.3",
        "symfony/expression-language": "<4.3",
        "symfony/security-http": "<4.3",
        "symfony/security-guard": "<4.3",
        "symfony/security-csrf": "<4.3",
        "symfony/security-core": "<4.3",
        "symfony/dependency-injection": "<4.3",
        "symfony/event-dispatcher": "<4.3",
        "symfony/http-kernel": "<4.3",
        "symfony/debug": "<4.3",
        "symfony/property-access": "<4.3",
        "symfony/property-info": "<4.3",
        "doctrine/dbal": "<2.9",
        "api-platform/core": "<2.4"
    }
}
```

And voilà:

* **with this PR**
    Memory usage: 193.33MiB (peak: 245.72MiB), time: 4.61s


I'm not saying that Composer should mention this anywhere as being best practice by all means! And I would also prefer to have a solution that works even better out of the box in the long term. But that's going to take some time and with this I could at least get my CI jobs running again on Travis. Because let's face it, the numbers here are still kind of okay but I'm having projects with 3-4 GB memory usage. Right now I'm just lost 🤷‍♂